+++ changelog.md
@@ -16,3 +16,4 @@
 - Add proper multiplayer weather system -Xeno333
 - Improve UI add new function to core_1042 -Xeno333
 - Add cooking and improve food API -Xeno333
+- Add aux1 uses via _1042_on_use -Xeno333